requirements.txt
streamlit~=1.45.1
python-dotenv~=1.1.0
langchain~=0.3.25
langchain-openai~=0.3.19
langchain-community~=0.3.24
python-docx~=1.1.2
PyMuPDF
scikit-learn~=1.6.1
pandas~=2.2.3

numpy~=2.2.6